pyproject.toml
[project]
name = "pi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-advanced-python-wrapper>=1.2.0",
    "babel>=2.17.0",
    "boto3>=1.38.23",
    "httpx>=0.28.1",
    "jupyterlab>=4.4.2",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "mysql-connector-python>=9.3.0",
    "pandas>=2.2.3",
    "psycopg>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "scikit-learn>=1.7.0",
    "sqlmodel>=0.0.24",
    "st-pages>=1.0.1",
    "streamlit>=1.45.1",
    "tqdm>=4.67.1",
]

[tool.ruff.lint]
select = [
    "A",
    "ANN",
    "ARG",
    "ASYNC",
    "B",
    # "BLE", # blind exception
    "C4",
    "C90",
    # "COM",
    "DOC",
    "DTZ",
    "E",
    "EM",
    "F",
    "FAST",
    # "FBT",
    "FIX",
    "FURB", # is unstable, search more later
    "I",
    "ISC",
    "N",
    "NPY",
    "PD",
    "PIE",
    "PERF",
    "PL",
    "PLE",
    "PLR",
    "PLW",
    "PT",
    "PTH",
    "Q",
    # "RET",
    "RUF",
    "SLF",
    "SIM",
    "TRY",
    "UP",
    "W",
]
preview = true
ignore = [
    "PLR0913",
    "PLR0917",
    "PD901",
    "SIM117",
    "RET501",
    "ANN401",
    "PERF401",
]

[tool.ruff.format]
preview = true
quote-style = "double"

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
line-length = 120
